refactor: JdbcTemplate과 SimpleJpa 분리
