+++ app/src/main/java/com/techcourse/dao/UserHistoryDao.java
@@ -4,6 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.interface21.jdbc.core.JdbcTemplate;
+import com.interface21.jdbc.core.jpa.SimpleJpa;
 import com.techcourse.domain.UserHistory;
 
 public class UserHistoryDao {
@@ -11,12 +12,14 @@ public class UserHistoryDao {
     private static final Logger log = LoggerFactory.getLogger(UserHistoryDao.class);
 
     private final JdbcTemplate jdbcTemplate;
+    private final SimpleJpa simpleJpa;
 
-    public UserHistoryDao(JdbcTemplate jdbcTemplate) {
+    public UserHistoryDao(final JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
+        this.simpleJpa = new SimpleJpa(jdbcTemplate);
     }
 
     public void log(final UserHistory userHistory) {
-        jdbcTemplate.insert(userHistory);
+        simpleJpa.insert(userHistory);
     }
 }
