+++ app/src/main/java/com/techcourse/domain/UserHistory.java
@@ -2,9 +2,9 @@ package com.techcourse.domain;
 
 import java.time.LocalDateTime;
 
-import com.interface21.jdbc.core.annotation.Column;
-import com.interface21.jdbc.core.annotation.Id;
-import com.interface21.jdbc.core.annotation.Table;
+import com.interface21.jdbc.core.jpa.annotation.Column;
+import com.interface21.jdbc.core.jpa.annotation.Id;
+import com.interface21.jdbc.core.jpa.annotation.Table;
 
 @Table(name = "user_history")
 public class UserHistory {
